pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool]

[tool.black]
line-length = 99

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.1.5"
tag_format = "$version"
version_files = [
    "tget/core/tget.py",
    "pyproject.toml"
]

[tool.poetry]
name = "tget"
version = "1.1.5"
description = "Search torrents from the command-line"
license = "MIT"
keywords = ["command", "line", "torrent"]
homepage = "https://github.com/rachmadaniHaryono/we-get"
authors = ["Levi Sabah <0xl3vi@gmail.com>"]
include = ["tget/txt/useragents.txt"]

[tool.poetry.dependencies]
python = ">=3.6.2"
beautifulsoup4 = "^4.10.0"
colorama = "^0.4.4"
docopt = "^0.6.2"
prompt-toolkit = ">=3.0.5"
pygments = ">=2.6.1"
requests = ">=2.27.1"

[tool.poetry.dev-dependencies]
pytest-flake8 = "^1.0.7"
vcrpy = "^4.1.1"

[tool.poetry.scripts]
tget = 'tget:main'
